Split out an opal-core library for tools
 * Fixes Issue #9869
 * Split `libopen-pal.so` into two libraries:
   - `libopen-pal_core.la` : Internal "core" portion of OPAL containing the essential source and MCA needed for mpicc/mpirun tools to link against. The "core" library is not installed.
   - `libopen-pal.la` : Includes "core" plus all of the other OPAL project sources. The `.so` version of this is installed.
 * The "core" library contains the following:
   - `opal/class`
   - `opal/mca/backtrace`
   - `opal/mca/dl`
   - `opal/mca/installdirs`
   - `opal/mca/threads`
   - `opal/mca/timer`
   - `opal/runtime/*_core.[c|h]`
   - `opal/runtime/opal_info_support.c`
   - `opal/util (most - see Makefile.am)`
   - `opal/util/keyval`
 * The "core" library is linked into the following tools instead of the full `libopen-pal.so`:
   - `ompi/tools/mpirun`
   - `ompi/tools/wrappers` (by extension of `opal/tools/wrappers`)
   - `opal/tools/wrappers`
 * The `opal/runtime` files were divided into a 'core' set vs 'main' set
