+++ opal/class/Makefile.am
@@ -3,7 +3,7 @@
 # Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
 #                         University Research and Technology
 #                         Corporation.  All rights reserved.
-# Copyright (c) 2004-2007 The University of Tennessee and The University
+# Copyright (c) 2004-2022 The University of Tennessee and The University
 #                         of Tennessee Research Foundation.  All rights
 #                         reserved.
 # Copyright (c) 2004-2005 High Performance Computing Center Stuttgart,
@@ -14,6 +14,7 @@
 # Copyright (c) 2014-2018 Los Alamos National Security, LLC. All rights
 #                         reserved.
 # Copyright (c) 2021      Nanook Consulting.  All rights reserved.
+# Copyright (c) 2022      Amazon.com, Inc. or its affiliates.  All Rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
@@ -42,7 +43,7 @@ headers += \
         class/opal_rb_tree.h \
         class/opal_interval_tree.h
 
-lib@OPAL_LIB_NAME@_la_SOURCES += \
+libopen_pal_core_la_SOURCES += \
         class/opal_bitmap.c \
         class/opal_cstring.c \
         class/opal_free_list.c \
