+++ opal/mca/installdirs/configure.m4
@@ -1,11 +1,15 @@
 dnl -*- shell-script -*-
 dnl
 dnl Copyright (c) 2006-2010 Sandia National Laboratories. All rights reserved.
+dnl Copyright (c) 2022      Amazon.com, Inc. or its affiliates.  All Rights reserved.
 dnl $COPYRIGHT$
 dnl
 dnl Additional copyrights may follow
 dnl
 dnl $HEADER$
 dnl
+
+dnl need in the core library
+AC_DEFUN([MCA_opal_installdirs_CORE_LIB], [1])
 
 AC_DEFUN([MCA_opal_installdirs_CONFIGURE_MODE], [PRIORITY])
